feat(editor-core): extract reusable editor schema
