+++ packages/editor-core/package.json
@@ -1,0 +1,56 @@
+{
+  "name": "@mdcms/editor-core",
+  "version": "0.1.0",
+  "description": "Reusable non-React editor schema and Markdown/MDX conversion primitives for MDCMS",
+  "type": "module",
+  "main": "./dist/index.js",
+  "module": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "files": [
+    "dist"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/mdcms-ai/mdcms.git",
+    "directory": "packages/editor-core"
+  },
+  "homepage": "https://github.com/mdcms-ai/mdcms#readme",
+  "bugs": "https://github.com/mdcms-ai/mdcms/issues",
+  "exports": {
+    "./package.json": "./package.json",
+    ".": {
+      "@mdcms/source": "./src/index.ts",
+      "bun": "./src/index.ts",
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.js",
+      "default": "./dist/index.js"
+    }
+  },
+  "nx": {
+    "name": "editor-core"
+  },
+  "scripts": {
+    "prepack": "node ../../scripts/strip-dev-export-conditions.js",
+    "postpack": "node ../../scripts/restore-package-json.js",
+    "test": "bun test ./src"
+  },
+  "dependencies": {
+    "@mdcms/shared": "^0.4.0",
+    "@tiptap/core": "^3.7.0",
+    "@tiptap/extension-code-block-lowlight": "^3.7.0",
+    "@tiptap/extension-highlight": "^3.7.0",
+    "@tiptap/extension-link": "^3.7.0",
+    "@tiptap/extension-task-item": "^3.7.0",
+    "@tiptap/extension-task-list": "^3.7.0",
+    "@tiptap/extension-underline": "^3.7.0",
+    "@tiptap/markdown": "^3.7.0",
+    "@tiptap/pm": "^3.7.0",
+    "@tiptap/starter-kit": "^3.7.0",
+    "lowlight": "^3.3.0",
+    "mdast-util-from-markdown": "^2.0.3",
+    "mdast-util-mdx": "^3.0.0",
+    "micromark-extension-mdx": "^2.0.0",
+    "tslib": "^2.3.0"
+  },
+  "license": "MIT"
+}
