+++ packages/editor-core/src/bun-test.d.ts
@@ -1,0 +1,3 @@
+declare module "bun:test" {
+  export function test(label: string, fn: () => void | Promise<void>): void;
+}
