+++ packages/editor-core/src/index.ts
@@ -1,0 +1,39 @@
+export {
+  createEditorCoreExtensions,
+  createEditorCoreLowlight,
+  type EditorCoreExtensionOptions,
+  type EditorCoreExtensionOverrides,
+} from "./editor-extensions.js";
+export {
+  extractMarkdownFromEditor,
+  parseMarkdownToDocument,
+  roundTripMarkdown,
+  serializeDocumentToMarkdown,
+} from "./markdown-pipeline.js";
+export {
+  MdxComponentExtension,
+  isMdxExpressionValue,
+  parseMdxAttributeValue,
+  parseMdxJsxAttributes,
+  serializeMdxJsxAttributes,
+  tokenizeMdxComponentBlock,
+  type MdxExpressionValue,
+} from "./mdx-component-extension.js";
+export {
+  MdxIntrinsicElementExtension,
+  tokenizeMdxIntrinsicElementBlock,
+} from "./mdx-intrinsic-element-extension.js";
+export { MdxIntrinsicInlineExtension } from "./mdx-intrinsic-inline-extension.js";
+export {
+  MDX_INTRINSIC_TEXT_BLOCK_ELEMENTS,
+  isMdxIntrinsicInlineName,
+  isMdxIntrinsicTextBlockName,
+} from "./mdx-intrinsic-inline.js";
+export {
+  MdxRawJsxExtension,
+  renderRawMdxJsxPreview,
+  tokenizeMdxRawJsxBlock,
+} from "./mdx-raw-jsx-extension.js";
+export { parseMdxMarkdownToTipTapDocument } from "./mdx-markdown-parser.js";
+export { StudioImageExtension } from "./studio-image-extension.js";
+export { HTML_VOID_ELEMENTS } from "./html-void-elements.js";
