+++ packages/studio/package.json
@@ -88,6 +88,7 @@
     "@fontsource-variable/geist-mono": "^5.2.7",
     "@fontsource-variable/inter": "^5.2.8",
     "@fontsource-variable/space-grotesk": "^5.2.10",
+    "@mdcms/editor-core": "^0.1.0",
     "@mdcms/shared": "^0.4.0",
     "@radix-ui/react-avatar": "1.1.11",
     "@radix-ui/react-dialog": "1.1.15",
