+++ packages/studio/src/lib/editor-extensions.ts
@@ -1,22 +1,20 @@
 import { Extension } from "@tiptap/core";
 import type { Extensions } from "@tiptap/core";
-import CodeBlockLowlight from "@tiptap/extension-code-block-lowlight";
-import Highlight from "@tiptap/extension-highlight";
-import Link from "@tiptap/extension-link";
-import TaskItem from "@tiptap/extension-task-item";
-import TaskList from "@tiptap/extension-task-list";
-import Underline from "@tiptap/extension-underline";
-import { Markdown } from "@tiptap/markdown";
+import type {} from "@tiptap/extension-code-block-lowlight";
+import type {} from "@tiptap/extension-highlight";
+import type {} from "@tiptap/extension-link";
+import type {} from "@tiptap/extension-task-item";
+import type {} from "@tiptap/extension-task-list";
+import type {} from "@tiptap/extension-underline";
+import type {} from "@tiptap/markdown";
 import { Plugin, PluginKey } from "@tiptap/pm/state";
 import { Decoration, DecorationSet } from "@tiptap/pm/view";
-import StarterKit from "@tiptap/starter-kit";
-import { common, createLowlight } from "lowlight";
-
-import { MdxComponentExtension } from "./mdx-component-extension.js";
-import { MdxIntrinsicElementExtension } from "./mdx-intrinsic-element-extension.js";
-import { MdxIntrinsicInlineExtension } from "./mdx-intrinsic-inline-extension.js";
-import { MdxRawJsxExtension } from "./mdx-raw-jsx-extension.js";
-import { StudioImageExtension } from "./studio-image-extension.js";
+import type {} from "@tiptap/starter-kit";
+import {
+  createEditorCoreExtensions,
+  createEditorCoreLowlight,
+  type EditorCoreExtensionOverrides,
+} from "@mdcms/editor-core";
 
 // Returns a lowlight instance seeded with the common language set and with
 // `highlightAuto` replaced by a plain-text no-op. CodeBlockLowlight falls
@@ -24,23 +22,7 @@ import { StudioImageExtension } from "./studio-image-extension.js";
 // default guesses at a grammar and renders tokens inside what the user
 // thinks is a plain-text block. Overriding it keeps "Plain text" honest
 // and matches the spec's "no auto-detection" decision.
-export function createStudioLowlight() {
-  const instance = createLowlight(common);
-
-  (instance as { highlightAuto: (value: string) => unknown }).highlightAuto = (
-    value: string,
-  ) => ({
-    type: "root",
-    data: { language: undefined },
-    children: [{ type: "text", value }],
-  });
-
-  return instance;
-}
-
-// Module-scope lowlight instance — language grammars are registered exactly
-// once for the lifetime of the process rather than per editor mount.
-const lowlightInstance = createStudioLowlight();
+export const createStudioLowlight = createEditorCoreLowlight;
 
 const BlurSelectionPreserver = Extension.create({
   name: "blurSelectionPreserver",
@@ -164,55 +146,15 @@ const EmptyParagraphHint = Extension.create({
   },
 });
 
-const HeadlessCodeBlock = CodeBlockLowlight.configure({
-  lowlight: lowlightInstance,
-  defaultLanguage: null,
-});
-
 // `createEditorExtensions` is safe to import from headless contexts (the
 // markdown pipeline, the CLI, document-editor.ts). The UI layer passes its
 // own React-wrapped `codeBlock` override so `@tiptap/react` only loads when
 // an editor is actually mounting in a browser.
-export function createEditorExtensions(options?: {
-  mdxRawJsx?: Extensions[number];
-  mdxComponent?: Extensions[number];
-  mdxIntrinsicElement?: Extensions[number];
-  image?: Extensions[number];
-  codeBlock?: Extensions[number];
-}): Extensions {
-  return [
-    StarterKit.configure({
-      codeBlock: false,
-      // The default dropcursor is a 1px `currentColor` line, which is
-      // almost invisible on the editor's typical white background and
-      // makes it hard to tell where a dragged MDX block will land. Use
-      // a thicker bar in the theme's primary accent so the drop target
-      // reads at a glance during reorder.
-      dropcursor: {
-        width: 4,
-        color: "var(--color-primary, #2563eb)",
-      },
-    }),
-    Underline,
-    Highlight,
-    BlurSelectionPreserver,
-    EmptyParagraphHint,
-    Link.configure({
-      openOnClick: false,
-      HTMLAttributes: {
-        rel: "noopener noreferrer nofollow",
-      },
-    }),
-    TaskList,
-    TaskItem.configure({
-      nested: true,
-    }),
-    options?.codeBlock ?? HeadlessCodeBlock,
-    options?.image ?? StudioImageExtension,
-    MdxIntrinsicInlineExtension,
-    options?.mdxComponent ?? MdxComponentExtension,
-    options?.mdxIntrinsicElement ?? MdxIntrinsicElementExtension,
-    options?.mdxRawJsx ?? MdxRawJsxExtension,
-    Markdown,
-  ];
+export function createEditorExtensions(
+  options?: EditorCoreExtensionOverrides,
+): Extensions {
+  return createEditorCoreExtensions({
+    ...options,
+    extensionsAfterHighlight: [BlurSelectionPreserver, EmptyParagraphHint],
+  });
 }
