+++ packages/studio/src/lib/html-void-elements.ts
@@ -1,16 +1,1 @@
-export const HTML_VOID_ELEMENTS = new Set([
-  "area",
-  "base",
-  "br",
-  "col",
-  "embed",
-  "hr",
-  "img",
-  "input",
-  "link",
-  "meta",
-  "param",
-  "source",
-  "track",
-  "wbr",
-]);
+export { HTML_VOID_ELEMENTS } from "@mdcms/editor-core";
