+++ packages/studio/tsconfig.lib.json
@@ -20,6 +20,9 @@
   "include": ["src/**/*.ts", "src/**/*.tsx"],
   "references": [
     {
+      "path": "../editor-core/tsconfig.lib.json"
+    },
+    {
       "path": "../shared/tsconfig.lib.json"
     }
   ]
